Update the VID and PID in mj64/config.h add the rev1/rev1.c add the rev2/rev2.c
